requirements.txt
Flask==2.2.3
flask-cors==3.0.10
opencv-python==4.7.0.72
torch==1.13.1
transformers==4.26.1
Pillow==9.4.0
numpy==1.23.5
